pyproject.toml
[tool.poetry]
name = "dnab"
version = "0.1.0"
description = "Api Annee Bissextile avec django ninja"
authors = ["Fanny Tutenuit <fanny.tutenuit@covage.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
django = "5.1.4"
django-ninja = "1.3.0"


[tool.poetry.group.dev.dependencies]
mypy = "1.14.1"
ruff = "0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
